refactor using new common parent class SparkAction
