+++ src/main/scala/io/smartdatalake/workflow/action/CustomSparkAction.scala
@@ -43,6 +43,7 @@ case class CustomSparkAction ( override val id: ActionObjectId,
                                override val breakDataFrameLineage: Boolean = false,
                                override val persist: Boolean = false,
                                override val initExecutionMode: Option[ExecutionMode] = None,
+                               override val executionMode: Option[ExecutionMode] = None,
                                override val metadata: Option[ActionMetadata] = None
 )(implicit instanceRegistry: InstanceRegistry) extends SparkSubFeedsAction {
